+++ oneview_redfish_toolkit/blueprints/event_service.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright (2017) Hewlett Packard Enterprise Development LP
+# Copyright (2017-2018) Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License"); you may
 # not use this file except in compliance with the License. You may obtain
@@ -16,19 +16,53 @@
 
 import logging
 
-
+from copy import deepcopy
 from flask import abort
 from flask import Blueprint
+from flask import request
 from flask import Response
 from flask_api import status
 
 from oneview_redfish_toolkit.api.errors import OneViewRedfishError
-from oneview_redfish_toolkit.api.event_service \
-    import EventService
+from oneview_redfish_toolkit.api.event import Event
+from oneview_redfish_toolkit.api.event_service import EventService
+
 from oneview_redfish_toolkit import util
 
 
 event_service = Blueprint("event_service", __name__)
+
+ONEVIEW_TEST_ALERT = {
+    "timestamp": "2018-02-12T20:12:03.231Z",
+    "resource": {
+        "category": "alerts",
+        "associatedResource": {
+            "resourceName": "0000A66101, bay 3",
+            "resourceUri": "/rest/server-hardware/"
+                           "30373737-3237-4D32-3230-313530314752"
+        }
+    }
+}
+
+ONEVIEW_TEST_TASK = {
+    "timestamp": "2018-02-12T20:12:03.231Z",
+    "resourceUri": "/rest/server-hardware/"
+                   "30373737-3237-4D32-3230-313530314752",
+    "changeType": None,
+    "resource": {
+        "category": "server-hardware",
+        "name": "0000A66101, bay 3"
+    }
+}
+
+REDFISH_TO_ONEVIEW_EVENTS = {
+    "ResourceAdded": "Created",
+    "ResourceUpdated": "Updated",
+    "ResourceRemoved": "Deleted",
+
+    # OneView does not have an equivalent change type to StatusChange
+    "StatusChange": "StatusChange"
+}
 
 
 @event_service.route("/redfish/v1/EventService/", methods=["GET"])
@@ -61,3 +95,60 @@ def get_event_service():
         # In case of error print exception and abort
         logging.exception(e)
         return abort(status.HTTP_500_INTERNAL_SERVER_ERROR)
+
+
+@event_service.route(
+    "/redfish/v1/EventService/Actions/EventService.SubmitTestEvent/",
+    methods=["POST"])
+def execute_test_event_action():
+    """Executes the SubmitTestEvent Action
+
+        Return a JSON containing the EventType received.
+        Logs exception of any error and return abort.
+
+        Returns:
+            JSON: JSON containing the EventType.
+
+        Exceptions:
+            OneViewRedfishError: When an invalid JSON is received.
+            return abort(400)
+
+            Exception: Unexpected error.
+            return abort(500)
+    """
+
+    try:
+        try:
+            event_type = request.get_json()['EventType']
+        except Exception:
+            raise OneViewRedfishError(
+                {'message': 'Invalid JSON data. Missing EventType property.'})
+
+        if (event_type not in util.subscriptions_by_type.keys()):
+            raise OneViewRedfishError(
+                {'message': 'Invalid EventType value: %s' % event_type})
+
+        # Creates a sample OneView SCMB message according to
+        # the value of 'event_type'
+        if (event_type == "Alert"):
+            message = deepcopy(ONEVIEW_TEST_ALERT)
+        else:
+            message = deepcopy(ONEVIEW_TEST_TASK)
+            message['changeType'] = REDFISH_TO_ONEVIEW_EVENTS[event_type]
+
+        event = Event(message)
+
+        util.dispatch_event(event)
+
+        json_str = event.serialize()
+
+        return Response(
+            response=json_str,
+            status=status.HTTP_200_OK,
+            mimetype='application/json')
+    except OneViewRedfishError as e:
+        logging.exception('Mapping error: {}'.format(e))
+        abort(status.HTTP_400_BAD_REQUEST, e.msg['message'])
+    except Exception as e:
+        logging.exception('Unexpected error: {}'.format(e))
+        abort(status.HTTP_500_INTERNAL_SERVER_ERROR)